SprintCapacityCalculator/requirements.txt
# Flask Web Framework
Flask==3.0.0
Werkzeug==3.1.3

# HTTP Requests
requests==2.31.0

# Environment Configuration
python-dotenv==1.0.0

# Core Dependencies (automatically installed with above packages)
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
urllib3==2.3.0
